chore: move interceptor to separate directory
