+++ src/main/kotlin/grpc/interceptor/AuthenticationInterceptor.kt
@@ -1,4 +1,4 @@
-package se.uulm.snowballr.backend.grpc
+package se.uulm.snowballr.backend.grpc.interceptor
 
 import io.github.oshai.kotlinlogging.KotlinLogging
 import io.grpc.Context
@@ -13,6 +13,7 @@ import io.jsonwebtoken.JwtException
 import se.uulm.snowballr.backend.auth.CookieUtils
 import se.uulm.snowballr.backend.auth.CookieUtils.parseCookies
 import se.uulm.snowballr.backend.auth.JwtUtils
+import se.uulm.snowballr.backend.grpc.GrpcContext
 import se.uulm.snowballr.backend.model.auth.AuthRequestState
 import se.uulm.snowballr.backend.model.jwt.ParsedJwtClaims
 import snowballr.SnowballRGrpcKt
